feat: add decimal/numeric support

+++ django_spanner/base.py
@@ -34,7 +34,7 @@ class DatabaseWrapper(BaseDatabaseWrapper):
         "CharField": "STRING(%(max_length)s)",
         "DateField": "DATE",
         "DateTimeField": "TIMESTAMP",
-        "DecimalField": "FLOAT64",
+        "DecimalField": "NUMERIC",
         "DurationField": "INT64",
         "EmailField": "STRING(%(max_length)s)",
         "FileField": "STRING(%(max_length)s)",
